+++ EngineSource/Engine/EngineProperties.h
@@ -16,7 +16,7 @@ namespace Project
 	void OnLaunch();
 	extern const char* ProjectName;
 }
-#define VERSION_STRING "1.10.0"
+#define VERSION_STRING "1.10.2"
 #define OPENGL_MIN_REQUIRED_VERSION "GL_VERSION_4_2"
 
 /**